draft script test
